backend/interview_project/requirements.txt
# Core dependencies
openai>=1.3.0
langchain>=0.0.285
langchain-openai>=0.0.2
pydantic>=2.4.2
transformers>=4.35.0
sentence-transformers>=2.2.2
keybert>=0.7.0
vaderSentiment>=3.3.2
textstat>=0.7.3
faiss-cpu>=1.7.4
numpy>=1.24.0
torch>=2.0.0
tqdm>=4.66.1

# Audio processing
sounddevice>=0.4.6
pygame>=2.5.2
boto3>=1.28.63
aiofiles>=23.2.1

# PDF generation
reportlab>=4.0.4

# Async utilities
asyncio>=3.4.3

# Optional: GPU support for faster embeddings
# faiss-gpu>=1.7.4
# For Apple Silicon Macs, use accelerated PyTorch
# torch>=2.0.0  # macOS builds install with pip install torch 